backend/requirements.txt
PyJWT>=2.10.0
# Core FastAPI and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
email-validator>=2.2.0
orjson>=3.10.0

# Database
sqlalchemy>=2.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
sqlmodel>=0.0.24

# HTTP Client and Networking
aiohttp>=3.11.0
aiohttp-socks>=0.10.0
python-socks>=2.7.0
PySocks>=1.7.1
aiofiles>=24.0.0
python-multipart>=0.0.20
aiodns>=3.5.0

# Email and SMTP/IMAP
aiosmtplib>=4.0.0
aioimaplib>=2.0.0

# Caching and Performance
async-lru>=2.0.0
tenacity>=9.0.0

# Background Tasks and Queue Management
celery>=5.5.3
redis>=6.2.0

# System Monitoring
psutil>=7.0.0

# Authentication and Security
passlib>=1.7.4
python-jose[cryptography]>=3.5.0
bcrypt>=4.3.0

# Natural Language Processing
nltk>=3.9.0

# HTML Processing
beautifulsoup4>=4.13.0

# File Processing
python-magic>=0.4.27

# Monitoring and Metrics
prometheus-client>=0.22.0

# Random String Generation
rstr>=3.2.0

# Additional Dependencies (Previously Missing)
cachetools>=6.1.0
httpx>=0.28.0
dkimpy>=1.1.0
lxml>=5.3.0
base58>=2.1.1
flower>=2.0.0
msgpack>=1.0.8

# Standard Libraries (usually included but listing for completeness)
requests>=2.31.0
python-dateutil>=2.9.0
typing-extensions>=4.13.0
cryptography>=45.0.0
